projects/sindhi-rag-translator/pyproject.toml
[project]
name = "sindhi-transliteration-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"  # <--- Change this line
dependencies = [
    "chainlit>=2.9.4",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]
